Refactor out HumanPoseAnalyzer; Add support for multiple HPAs
